+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 Changes
 -------
 v1.6.0 (??.1?.2020)
- * Moving table classes to PSR-4
+ * #59 Moving table classes to PSR-4
+ * #71 ignore matches of teams out of competitions in table calculation
+ * fix some PHP issues
 
 v1.5.2 (14.06.2020)
  * labels in plugin flexforms fixed